requirements.txt
fastapi
uvicorn[standard]
requests
pytz
mlflow
pandas
torch
transformers
datasets
accelerate
seqeval
sentencepiece
scikit-learn
tqdm
protobuf
numpy
google-cloud-bigquery
google-auth
google-api-core
